requirements.txt
# Core Django and REST Framework
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.10

# CORS headers for frontend communication
django-cors-headers==4.6.0

# Environment variables
python-decouple==3.8

# Image processing for solution photos
Pillow>=10.0.0

# Google Gemini API
google-generativeai==0.8.3

# Development tools
django-extensions==3.2.3

# Production server
gunicorn==21.2.0
whitenoise==6.7.0
